work in progress to refactor, beginning with scheduling pages
