+++ src/pages/About.jsx
@@ -3,6 +3,6 @@ import React from 'react'
 
 export default function About() {
   return (
-    <div>About</div>
+    <div>About the site</div>
   )
 }
