+++ src/pages/GameInfo/gameInfo.css
@@ -33,10 +33,10 @@
 /* .scorecards {
   display: flex;
 } */
-
-.away-scorecard, .home-scorecard {
+/* 
+.scorecard {
   display: flex;
-}
+} */
 
 .player-columns-categories th {
   height: 1.3rem;
@@ -52,4 +52,20 @@
 
 table, th, td, tr {
   border: solid black 1px;
+}
+
+.scorecard-cell {
+  color:white;
+}
+
+.scoring-cell {
+  background-color: pink;
+}
+
+.empty-cell {
+  background-color: darkkhaki;
+}
+
+.scorecard-inning-header {
+  width: 3rem;
 }
