+++ src/pages/Scheduling/SeasonSchedule/components/Error.jsx
@@ -1,3 +1,4 @@
+// TODO: lift this into a shared error message/component so other pages can use it
 import React from "react";
 
 export default function Error() {
