+++ src/pages/Scheduling/SeasonSchedule/components/SearchResultsString.jsx
