+++ src/pages/Scheduling/components/Game/game.css
