+++ src/pages/Scheduling/index.js
@@ -1,0 +1,4 @@
+import { TodaysSchedule } from "./TodaysSchedule/TodaysSchedule";
+import { SeasonSchedule } from "./SeasonSchedule/SeasonSchedule";
+
+export { TodaysSchedule, SeasonSchedule };
